requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
celery[redis]==5.4.0
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.1
httpx==0.28.1
boto3==1.35.86
# Pipeline (feat/pipeline)
python-multipart==0.0.20
python-pptx==1.0.2
Pillow==11.1.0
anthropic==0.43.0
openai==1.58.1
deepl==1.18.0
google-cloud-translate==3.18.0
# HeyGen TTS (feat/heygen)
google-cloud-texttospeech==2.21.1
# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
